replace mysql on sidecar step by step
