+++ sidecar/src/db.rs
@@ -135,82 +135,3 @@ pub async fn query_pending_orders(
         .map(|o| (symbol.clone(), o).into())
         .collect::<Vec<_>>())
 }
-
-// TODO save the relayer
-pub async fn save_trading_command(
-    pool: &Pool<MySql>,
-    user_id: impl ToString,
-    cmd: TradingCommand,
-    relayer: impl ToString,
-) -> anyhow::Result<u64> {
-    // TODO
-    let fix_cmd_signature = "169d796416023558ef5c2580ef38c1c4f43f3c06f76ceab2412e6fc5d486a36eb0a9cb808dd4eb72f6264b4113c1a722479be205edc84d6ac5403d33d09b0087";
-    let fix_cmd_nonce = 40020u32;
-    let direction = cmd.get_direction_if_trade();
-    match cmd {
-        TradingCommand::Cancel {
-            base,
-            quote,
-            order_id,
-        } => {
-            let mut cancel = Command::default();
-            cancel.order_id = Some(order_id);
-            cancel.base = Some(base);
-            cancel.cmd = CANCEL;
-            cancel.quote = Some(quote);
-            cancel.user_id = Some(user_id.to_string());
-            cancel.signature = Some(fix_cmd_signature.to_string());
-            cancel.nonce = Some(fix_cmd_nonce);
-            sqlx::query("insert into t_sequence(f_cmd) values(?)")
-                .bind(serde_json::to_string(&cancel).expect("jsonser;qed"))
-                .execute(pool)
-                .await?;
-            Ok(order_id)
-        }
-        TradingCommand::Ask {
-            base,
-            quote,
-            amount,
-            price,
-        }
-        | TradingCommand::Bid {
-            base,
-            quote,
-            amount,
-            price,
-        } => {
-            let mut tx = pool.begin().await?;
-            let sql = format!("insert into t_order_{}_{}(f_user_id,f_amount,f_price,f_order_type) values(?,?,?,?)",
-                base, quote
-            );
-            sqlx::query(sql.as_str())
-                .bind(user_id.to_string())
-                .bind(amount.clone())
-                .bind(price.clone())
-                .bind(direction)
-                .execute(&mut tx)
-                .await?;
-            let result = sqlx::query("select LAST_INSERT_ID() as id")
-                .fetch_one(&mut tx)
-                .await?;
-            let id: u64 = result.get("id");
-            let mut place = Command::default();
-            place.cmd = direction.expect("ask_or_bid;qed").into();
-            place.order_id = Some(id);
-            place.base = Some(base);
-            place.quote = Some(quote);
-            place.signature = Some(fix_cmd_signature.to_string());
-            place.user_id = Some(user_id.to_string());
-            place.price = Decimal::from_str(&price).ok();
-            place.amount = Decimal::from_str(&amount).ok();
-            place.nonce = Some(fix_cmd_nonce);
-            place.broker = Some(relayer.to_string());
-            sqlx::query("insert into t_sequence(f_cmd) values(?)")
-                .bind(serde_json::to_string(&place).expect("jsonser;qed"))
-                .execute(&mut tx)
-                .await?;
-            tx.commit().await?;
-            Ok(id)
-        }
-    }
-}
